+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added *AutoCrop* position marker type
 - Added /about (GET) for health checks and monitoring
 - Added /shutdown (POST) for graceful shutdown
+- Added /tracks/purge (POST) to purge track sources and tracks by URI
 - Print socket address to *stdout* for connecting clients through an ephemeral port
 
 ### Changed